Included URL of Unified Hosts in version check
The script is aware of the current Unified Hosts option you prefer, such
as by itself or with which extensions. This also enables the script to
be smarter about updates to know that you should update if your current
version is older or if the URL you currently want is different from the
one you have.