Some minor improvements
Reorganized functions to make main script more easy to follow for
readability sake. Added the option to open your hosts file in notepad
before exiting. Updated the readme to specify downloading the zip
package and not just the batch file by itself and also posted a request
for more feedback.